\name{SuperLearner-package}
\alias{SuperLearner-package}

\docType{package}
\title{
Super Learner prediction function and library.
}
\description{
This package implements the Super Learner prediction method and contains a library of prediction algorithms to be used in the super learner.
}
\details{
\tabular{ll}{
Package: \tab SuperLearner\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2009-09-28\cr
License: \tab GPLv3\cr
LazyLoad: \tab yes\cr
}
}
\author{
Eric C. Polley

Maintainer: Eric C Polley <ecpolley@berkeley.edu>
}
\references{
	van der Laan, M. J., Polley, E. C. and Hubbard, A. E. (2008) Super Learner, \emph{Statistical Applications of Genetics and Molecular Biology}, \bold{6}, article 25. \url{http://www.bepress.com/sagmb/vol6/iss1/art25}  
}
\seealso{
 \code{\link{SuperLearner}}
}

